Refactored SensorData Model filter advanced query
